refactor 'check search bundle' service task implementation
